LPS-139540 Moves code to WidgetTemplatesTemplateDisplayContext
